requirements.txt
# pip install -r requirements.txt
slack-bolt
langflow
python-dotenv